pyproject.toml
[project]
name = "console-rpg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.5.0",
    "python-dotenv>=1.0.1",
]
